Backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1

# HTTP clients
httpx==0.27.0
requests==2.31.0

# Code parsing
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.0
tree-sitter-typescript==0.21.0

# Embeddings (local, no GPU needed)
sentence-transformers==2.7.0
torch==2.2.2 --index-url https://download.pytorch.org/whl/cpu

# Vector DB client
qdrant-client==1.9.1

# Keyword search
rank-bm25==0.2.2

# GitHub API
PyGithub==2.3.0

# Code analysis / linting
semgrep==1.72.0

# Async support
asyncio==3.4.3

# Utilities
numpy==1.26.4
tqdm==4.66.4
